refactor: use SparqlUpdateWriter graphIri option instead of wrapper class
Replace ValidationGraphWriter with the new graphIri callback on
SparqlUpdateWriter (added in ldelements/lde#409, shipped as
@lde/pipeline@0.30.1). The dataset's IRI never changes — only where the
quads land — so the wrapper-rewrites-dataset.iri dance was conflating
dataset identity with output location.

What remains is the policy: validationGraphIri(datasetIri), unchanged.
The wrapper class is gone; main.ts wires the policy directly:

  new SparqlUpdateWriter({
    endpoint,
    auth,
    graphIri: dataset => validationGraphIri(dataset.iri),
  });

Net: 53 lines deleted, no behavioural change.
