+++ package.json
@@ -32,11 +32,11 @@
   "dependencies": {
     "@lde/dataset": "^0.7.4",
     "@lde/dataset-registry-client": "^0.8.0",
-    "@lde/pipeline": "^0.30.0",
-    "@lde/pipeline-console-reporter": "^0.21.0",
-    "@lde/pipeline-shacl-sampler": "^0.4.0",
-    "@lde/pipeline-shacl-validator": "^0.12.0",
-    "@lde/pipeline-void": "^0.28.0",
+    "@lde/pipeline": "^0.30.1",
+    "@lde/pipeline-console-reporter": "^0.21.1",
+    "@lde/pipeline-shacl-sampler": "^0.4.1",
+    "@lde/pipeline-shacl-validator": "^0.12.1",
+    "@lde/pipeline-void": "^0.28.1",
     "@lde/sparql-qlever": "^0.14.4",
     "@netwerk-digitaal-erfgoed/network-of-terms-catalog": "^10.19.22",
     "env-schema": "^7.0.0",
